[Fix] #51 QuizWrongView,ViewController , QuizCompleteView,ViewController 코드규칙에 맞춰서 수정
