+++ FindAnimalFriends/FindAnimalFriends/ViewController/QuizWrongViewController.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 import UIKit
-class QuizWrongViewController: UIViewController{
+class QuizWrongViewController: UIViewController {
     
     private var quizWrongView : QuizWrongView?
     
@@ -15,10 +15,10 @@ class QuizWrongViewController: UIViewController{
         super.viewDidLoad()
         view.backgroundColor = .systemBackground
         quizWrongView = QuizWrongView(frame: CGRect(x: 0, y: 0, width: view.bounds.width, height: view.bounds.height))
-        quizWrongView!.completeButton.addTarget(self, action: #selector(CloseCollectView), for: .touchUpInside)
+        quizWrongView!.completeButton.addTarget(self, action: #selector(closeQuizWrongView), for: .touchUpInside)
         view.addSubview(quizWrongView!)
     }
-    @objc func CloseCollectView(){
+    @objc func closeQuizWrongView() {
         quizWrongView!.removeFromSuperview()
     }
